rearranged functions into relevant classes as a first step
